+++ CHANGELOG.md
@@ -67,6 +67,7 @@ All notable changes to this project will be documented in this file.
 - Fixed module import error by adding .js extension to @google-cloud/speech/build/protos/protos import
 - Fixed CommonJS module import error by using default import for google from @google-cloud/speech/build/protos/protos.js
 - Fixed "Invalid hook call" error in TranscriptDetails by correctly using the useSettings hook
+- Fixed Swift Recorder issues by enabling software recording mode by default
 
 ## [0.3.1] - 2024-05-21